pyproject.toml
[project]
name = "space"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "claude-agent-sdk>=0.1.58",
    "python-dotenv>=1.2.2",
    "requests>=2.33.1",
]
